docs: show how to use optional options in the schema example
I am using the schema validation and was looking for `nullable` solution. I found in the doc https://express-validator.github.io/docs/validation-chain-api.html that `optional` has the `nullable` option but I did not know it should be under `options`. It took me for a while to figure it out. So, I think updating this doc might help other people in future.